fix: PositiveNum 프로퍼티 타입 수정
- 숫자형 타입에 대한 검증을 extractor 에서 수행

# step 1

fix: PositiveNum 프로퍼티 타입 수정

- 숫자형 타입에 대한 검증을 extractor 에서 수행

# step 1

+++ src/main/kotlin/calculator/dto/PositiveNums.kt
@@ -2,6 +2,6 @@ package calculator.dto
 
 data class PositiveNums(val positiveNums: List<PositiveNum>) {
     fun sum(): Int {
-        return positiveNums.sumOf { it.value.toInt() }
+        return positiveNums.sumOf { it.value }
     }
 }
